requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
deprecation==2.1.0
Django==4.1.1
django-axes==5.39.0
django-cors-headers==3.13.0
django-csp==3.7
django-filter==22.1
django-ipware==4.0.2
django-user-sessions==1.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
joblib==1.2.0
lxml==4.9.1
Markdown==3.4.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pikepdf==6.0.2
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0